refactor: Moved tls from header to source code
